A Deep Learning System for Detection of Real and Fake Human Face
The widespread use of face recognition system has led a multiple attempts to fool the systems in order to impersonate another person .It is like filtering step of the existing system. There is a wide range of fake attacks, and the methods used to combat one type of attack may not be effective against other attacks. Our  method to detect fake facial images based on deep neural networks. The   use of deep learning improved the true acceptance rate and significantly reduced the error values.